Added mappings for passengers and flights

+++ src/main/java/com/example/spring_cinema_lab/controllers/MovieController.java
@@ -31,7 +31,7 @@ public class MovieController {
     }
 
     @GetMapping (value = "/{id}")
-    public ResponseEntity<Movie> getMovieById(@PathVariable int id) {
+    public ResponseEntity<Movie> getMovieById(@PathVariable long id) {
         Optional<Movie> movie = movieService.getMovieById(id);
         if (movie.isPresent()) {
             return new ResponseEntity<>(movie.get(), HttpStatus.OK);
